softeng24-13/requirements.txt
Flask
mysql-connector-python
pandas
SQLAlchemy
requests
